[ckks] : the API used by the bootstrapping to encode and evaluate diagonalized matrices is now public.
[encoder] : new API -> EncodeDiagMatrixAtLvl
[evaluator] : new API -> MultiplyByDiagMatrice

Removed the memory pool of the bootstrapper that was used in the matrix x vector product and transfered it into the evaluator.
