+++ ckks/bootstrapp_test.go
@@ -43,7 +43,7 @@ func TestBootstrapp(t *testing.T) {
 		// Insecure params for fast testing only
 		if !*flagLongTest {
 			params.logN = 14
-			params.logSlots = 13
+			params.logSlots = 12
 		}
 
 		if testContext, err = genTestParams(params, btpParams.H); err != nil {
